Do not specify content-length and content-encoding
The size reported by GCS is the compressed size, but the storage.Reader
decompresses on the fly causing a mismatch and subsequent truncation or
crash.

Setting content-encoding to the content-encoding of the file in the bucket
did not make much sense in the first place, let's just let the server
determine whatever it likes. We could pass-through gzipped data if the
file in the bucket is gzipped and the client accepts gzip, but that is
a lot of effort for very little benefit.

The status code is set to 200 by default, so get rid of that as well.